Add methods for creating person, adding book, adding rentals & refactor from modules to class
